+++ it/server/src/test/java/com/walmartlabs/concord/it/server/PolicyIT.java
@@ -1,0 +1,91 @@
+package com.walmartlabs.concord.it.server;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import com.walmartlabs.concord.client.*;
+import org.junit.Test;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static com.walmartlabs.concord.it.common.ITUtils.archive;
+import static com.walmartlabs.concord.it.common.ServerClient.assertLog;
+import static com.walmartlabs.concord.it.common.ServerClient.waitForCompletion;
+import static java.util.Collections.singletonMap;
+
+public class PolicyIT extends AbstractServerIT {
+
+    @Test(timeout = 60000)
+    public void testCfg() throws Exception {
+        String orgName = "org_" + randomString();
+        OrganizationsApi organizationsApi = new OrganizationsApi(getApiClient());
+        organizationsApi.createOrUpdate(new OrganizationEntry().setName(orgName));
+
+        String projectName = "project_" + randomString();
+        ProjectsApi projectsApi = new ProjectsApi(getApiClient());
+        projectsApi.createOrUpdate(orgName, new ProjectEntry()
+                .setName(projectName)
+                .setAcceptsRawPayload(true));
+
+        String policyName = "policy_" + randomString();
+        PolicyApi policyApi = new PolicyApi(getApiClient());
+        policyApi.createOrUpdate(new PolicyEntry()
+                .setName(policyName)
+                .setRules(singletonMap("processCfg",
+                        singletonMap("arguments",
+                                singletonMap("x",
+                                        singletonMap("name", "Concord"))))));
+
+        policyApi.link(policyName, new PolicyLinkEntry()
+                .setOrgName(orgName)
+                .setProjectName(projectName));
+
+        // ---
+
+        byte[] payload = archive(ProcessIT.class.getResource("policyCfg").toURI());
+
+        StartProcessResponse spr = start(payload);
+
+        ProcessApi processApi = new ProcessApi(getApiClient());
+        ProcessEntry pir = waitForCompletion(processApi, spr.getInstanceId());
+
+        // ---
+
+        byte[] ab = getLog(pir.getLogFileName());
+        assertLog(".*Hello, Stranger.*", ab);
+
+        // ---
+
+        Map<String, Object> input = new HashMap<>();
+        input.put("archive", payload);
+        input.put("org", orgName);
+        input.put("project", projectName);
+
+        spr = start(input);
+
+        pir = waitForCompletion(processApi, spr.getInstanceId());
+
+        // ---
+
+        ab = getLog(pir.getLogFileName());
+        assertLog(".*Hello, Concord.*", ab);
+    }
+}