+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/PolicyReader.java
@@ -1,0 +1,60 @@
+package com.walmartlabs.concord.server.process;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.walmartlabs.concord.project.InternalConstants;
+
+import javax.inject.Named;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Collections;
+import java.util.Map;
+import java.util.UUID;
+
+@Named
+public class PolicyReader {
+
+    private final ObjectMapper objectMapper;
+
+    public PolicyReader() {
+        this.objectMapper = new ObjectMapper();
+    }
+
+    public  Map<String, Object> readPolicy(Payload payload) {
+        return readPolicy(payload.getInstanceId(), payload.getHeader(Payload.WORKSPACE_DIR));
+    }
+
+    @SuppressWarnings("unchecked")
+    public Map<String, Object> readPolicy(UUID instanceId, Path workDir) {
+        Path policyFile = workDir.resolve(InternalConstants.Files.CONCORD_SYSTEM_DIR_NAME).resolve(InternalConstants.Files.POLICY_FILE_NAME);
+        if (!Files.exists(policyFile)) {
+            return Collections.emptyMap();
+        }
+
+        try {
+            return objectMapper.readValue(policyFile.toFile(), Map.class);
+        } catch (IOException e) {
+            throw new ProcessException(instanceId, "Reading process policy error", e);
+        }
+    }
+}